set up main.js to import modules
