+++ index.html
@@ -6,6 +6,7 @@
   <meta http-equiv="X-UA-Compatible" content="IE=edge">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <link rel="stylesheet" href="index.css" />
+  <script type="module" src="index.js"></script>
   <title>Awesome Books</title>
 </head>
 
@@ -69,9 +70,6 @@
     </footer>
 
   </div> <!-- END OF GRID-CONTAINER -->
-
-
-  <script src="index.js"></script>
 
 </body>
