+++ modules/book.js
@@ -1,0 +1,5 @@
+export default class Book {
+  title;
+
+  author;
+}
